+++ include/RNQuaternion.h
@@ -5,41 +5,41 @@
 
 class RNQuaternion{
 private:
-	float scalar;
+	double scalar;
 	RNVector3 axis;
 public:
-	RNQuaternion(const float scalar = 0, const RNVector3 axis = RNVector3());
-	RNQuaternion (const float scalar, const float a, const float b,const float c);
-	static RNQuaternion initFromEuler(const float yaw, const float pitch, const float roll);
-	static RNQuaternion initFromRPY(const float roll, const float pitch, const float yaw);
+	RNQuaternion(const double scalar = 0, const RNVector3 axis = RNVector3());
+	RNQuaternion (const double scalar, const double a, const double b,const double c);
+	static RNQuaternion initFromEuler(const double yaw, const double pitch, const double roll);
+	static RNQuaternion initFromRPY(const double roll, const double pitch, const double yaw);
 	const char* toString() const;
 
 	void operator=(const RNQuaternion& v);
 
 	RNQuaternion operator+(const RNQuaternion& v);
-	RNQuaternion operator+(const float& v);
+	RNQuaternion operator+(const double& v);
 	friend RNQuaternion operator+(const RNQuaternion& v1, const RNQuaternion& v2);
 
 	RNQuaternion& operator+=(const RNQuaternion& v);
-	RNQuaternion& operator+=(const float& v);
+	RNQuaternion& operator+=(const double& v);
 
 	RNQuaternion operator-() const;
 	RNQuaternion operator-(const RNQuaternion& v);
-	RNQuaternion operator-(const float& v);
+	RNQuaternion operator-(const double& v);
 	friend RNQuaternion operator-(const RNQuaternion& v1, const RNQuaternion& v2);
 
 	RNQuaternion& operator-=(const RNQuaternion& v);
-	RNQuaternion& operator-=(const float& v);
+	RNQuaternion& operator-=(const double& v);
 
 	RNQuaternion operator*(const RNQuaternion& v);
-	RNQuaternion operator*(const float& v);
+	RNQuaternion operator*(const double& v);
 	friend RNQuaternion operator*(const RNQuaternion& v1, const RNQuaternion& v2);
 
-	RNQuaternion& operator*=(const float& v);
+	RNQuaternion& operator*=(const double& v);
 
-	RNQuaternion operator/(const float& v);
+	RNQuaternion operator/(const double& v);
 
-	RNQuaternion& operator/=(const float& v);
+	RNQuaternion& operator/=(const double& v);
 
 	friend RNQuaternion operator!(RNQuaternion m);
 
@@ -48,36 +48,36 @@ public:
 
 	bool isZero();
 
-	float getScalar() const;
+	double getScalar() const;
 	RNVector3 getAxis() const;
-	float getX() const;
-	float getY() const;
-	float getZ() const;
+	double getX() const;
+	double getY() const;
+	double getZ() const;
 
-	float dot(const RNQuaternion& v) const;
-	float length() const;
-	float length2() const;
+	double dot(const RNQuaternion& v) const;
+	double length() const;
+	double length2() const;
 
-	float angleBetween(const RNQuaternion& q) const;
-	float angleShortestPath(const RNQuaternion& q) const;
+	double angleBetween(const RNQuaternion& q) const;
+	double angleShortestPath(const RNQuaternion& q) const;
 
 	RNQuaternion exp();
 	RNQuaternion ln();
-	RNQuaternion pow(float n);
+	RNQuaternion pow(double n);
 	RNQuaternion sign();
 
 	RNQuaternion norm();
 	RNQuaternion conjugate();
 	RNQuaternion inverse();
-	RNQuaternion slerp(const RNQuaternion& q, const float& t) const;
+	RNQuaternion slerp(const RNQuaternion& q, const double& t) const;
 
 	static const RNQuaternion& eye();
 
-	void setScalar(float scalar);
+	void setScalar(double scalar);
 	void setAxis(RNVector3 axis);
 
-	RNQuaternion slerpQuat (const RNQuaternion& q, const float& t);
-	float toAngle();
+	RNQuaternion slerpQuat (const RNQuaternion& q, const double& t);
+	double toAngle();
 	
 };
