+++ include/fl/fuzzyStats.h
@@ -8,21 +8,21 @@ namespace fuzzy{
 
 	class FStats{
 	private:
-		static float area(float x1, float x2, float x3, float x4);
+		static double area(double x1, double x2, double x3, double x4);
 	public:
 		
-		static float evaluateMF(fl::Term* member, float value);
-		static std::vector<float> evaluateMF(fl::Term* member, std::vector<float> values);
+		static double evaluateMF(fl::Term* member, double value);
+		static std::vector<double> evaluateMF(fl::Term* member, std::vector<double> values);
 		
-		static float expectation(float x1, float x2, float x3, float x4);
-		static float expectation(std::vector<float> evaluatedMembership, std::vector<float> values);
+		static double expectation(double x1, double x2, double x3, double x4);
+		static double expectation(std::vector<double> evaluatedMembership, std::vector<double> values);
 
-		static float uncertainty(float x1, float x2, float x3, float x4);
-		static float uncertainty(std::vector<float> evaluatedMembership, std::vector<float> values);
+		static double uncertainty(double x1, double x2, double x3, double x4);
+		static double uncertainty(std::vector<double> evaluatedMembership, std::vector<double> values);
 		
-		static float dependency(std::vector<float> evaluatedMembershipX, std::vector<float> x, std::vector<float> evaluatedMembershipY, std::vector<float> y);
-		static float standarDeviation(std::vector<float> evaluatedMembership, std::vector<float> values);
-		static float correlationCoefficient(std::vector<float> evaluatedMembershipX, std::vector<float> x, std::vector<float> evaluatedMembershipY, std::vector<float> y);
+		static double dependency(std::vector<double> evaluatedMembershipX, std::vector<double> x, std::vector<double> evaluatedMembershipY, std::vector<double> y);
+		static double standarDeviation(std::vector<double> evaluatedMembership, std::vector<double> values);
+		static double correlationCoefficient(std::vector<double> evaluatedMembershipX, std::vector<double> x, std::vector<double> evaluatedMembershipY, std::vector<double> y);
 	};
 }
 #endif
