+++ include/stats.h
@@ -10,23 +10,23 @@
 
 class stats{
 public:
-	static float max(std::vector<float> values);
-	static float min(std::vector<float> values);
+	static double max(std::vector<double> values);
+	static double min(std::vector<double> values);
 	
-	static float sum(std::vector<float> values);
-	static std::vector<int> findIndicesEqual(std::vector<float> values, float k);
-	static std::vector<int> findIndicesHigherThan(std::vector<float> values, float k);
-	static std::vector<int> findIndicesLessThan(std::vector<float> values, float k);
+	static double sum(std::vector<double> values);
+	static std::vector<int> findIndicesEqual(std::vector<double> values, double k);
+	static std::vector<int> findIndicesHigherThan(std::vector<double> values, double k);
+	static std::vector<int> findIndicesLessThan(std::vector<double> values, double k);
         
-	static bool isInf(float value);
-	static bool isNaN(float value);
+	static bool isInf(double value);
+	static bool isNaN(double value);
 	
-	static float expectation(std::vector<float> values);
-	static float variance(std::vector<float> values);
-	static float variance(std::vector<float> values, float expectation);
-	static float covariance(std::vector<float> x, std::vector<float> y);
-	static float standarDeviation(std::vector<float> values);
-	static float correlationCoefficient(std::vector<float> x, std::vector<float> y);
+	static double expectation(std::vector<double> values);
+	static double variance(std::vector<double> values);
+	static double variance(std::vector<double> values, double expectation);
+	static double covariance(std::vector<double> x, std::vector<double> y);
+	static double standarDeviation(std::vector<double> values);
+	static double correlationCoefficient(std::vector<double> x, std::vector<double> y);
 };
 
 #endif
